Add moe-bot package and module, enable it with our config
For updating the package's dependencies, a workaround can be found in
moe-bot/package.nix because this update script was made for nixpkgs.

Another workaround is being used for managing secrets. Currently the
secret is not stored in git and it is stored in plain text in the nix
store. For more information see rebuild.sh