Add localhost server location to Nginx This fixes an issue with gepDrive where authentication headers wouldn't work.
This is required due to the way Nginx resolves server names and handles requests.
When one accesses the server using a domain name (e.g., example.com or some-service.example.com),
Nginx uses the "Host" header in the HTTP requestto determine which server block to use.
However, when one accesses the server using a private IP or localhost,
the "Host" header might not be present or might be different.
In such cases, Nginx might default to the first server block defined
if it cannot match the "Host" header to any specific server block.